# gateway/api-doc.yaml
openapi: "3.0.0"
info:
  title: Oulipo API overview
  version: 0.1.0
security:
  - BasicAuth:
      - user
paths:
...
  /api/accounts:
    get:
      summary: Get all accounts
      operationId: getAllAccounts
      tags:
        - accounts
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        '200':
          description: Get all accounts success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Accounts"
  /api/accounts/{accountId}:
    get:
      summary: Get an account
      operationId: getAccount
      tags:
        - accounts
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Get account success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        '404':
          description:  Account not found by that id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
  parameters:
    offset:
      name: offset
      in: query
      description: How many ordered resources to skip before
      required: false
      schema:
        type: integer
        format: int32
    limit:
      name: limit
      in: query
      description: How many ordered resources to return in one request
      required: false
      schema:
        type: integer
        format: int32
  schemas:
    ApiValidationError:
      required:
        - field;
        - rejectedValue;
        - message;
      properties:
        id:
          type: integer
          format: int64
        rejectedValue:
          type: string
        message:
          type: string
...
    Account:
      required:
        - username
        - email
        - hashedPassword
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        hashedPassword:
          type: string
    Accounts:
      type: array
      items:
        $ref: "#/components/schemas/Account"
    Error:
      required:
        - status
        - timestamp
        - message
        - debugMessage
        - subErrors
      properties:
        status:
          type: string
        timestamps:
          type: string
        message:
          type: string
        debugMessage:
          type: array
          items:
            $ref: "#/components/schemas/ApiValidationError"


  #GET /api/users/?filterId=1234-abcd
  #It seems that resource filtering/searching can be implemented in a RESTful way. The idea is to introduce a new endpoint called /filters/ or /api/filters/.
  #
  #Using this endpoint filter can be considered as a resource and hence created via POST method. This way - of course - body can be used to carry all the parameters as well as complex search/filter structures can be created.